Upd: ComparerExpression`1 class moved from a ComparerBuilder<>. Upd: ComparerBuilder class added.
