+++ ComparerBuilder/ComparerBuilder.csproj
@@ -41,11 +41,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ComparerBuilder.cs" />
     <Compile Include="ComparerBuilder`1.cs" />
+    <Compile Include="ComparerExpression`1.cs" />
     <Compile Include="Comparers.cs" />
     <Compile Include="DefaultInterception.cs" />
     <Compile Include="IComparerExpression.cs" />
     <Compile Include="ComparerBuilderInterception.cs" />
+    <Compile Include="NestedComparerExpression`1.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReplaceVisitor.cs" />
     <Compile Include="SourceInfo.cs" />
